It's a lot of just adding content, less about NextJS Topics
