+++ 25 Foodies Starting Project/app/globals.css
@@ -1,57 +1,42 @@
 @import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@400;600;700&family=Montserrat:wght@400;700&display=swap');
 
 * {
-  box-sizing: border-box;
+	box-sizing: border-box;
 }
 
 html {
-  font-family: 'Quicksand', sans-serif;
+	font-family: 'Quicksand', sans-serif;
 }
 
 body {
-  margin: 0;
-  background: radial-gradient(#282c34, #282c34);
-}
-
-.header-background {
-  position: absolute;
-  width: 100%;
-  height: 320px;
-  top: 0;
-  left: 0;
-  z-index: -1;
-}
-
-svg {
-  display: block;
-  width: 100%;
-  height: auto;
+	margin: 0;
+	background: radial-gradient(#282c34, #282c34);
 }
 
 .not-found,
 .error {
-  margin-top: 5rem;
-  text-align: center;
+	margin-top: 5rem;
+	text-align: center;
 }
 
 .not-found h1,
 .error h1 {
-  font-size: 5rem;
-  margin: 0px;
-  font-weight: 900;
-  color: #262626;
-  text-transform: uppercase;
-  background: linear-gradient(90deg, #f9572a, #ffc905);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-size: cover;
-  background-position: center;
-  font-family: 'Montserrat', sans-serif;
+	font-size: 5rem;
+	margin: 0px;
+	font-weight: 900;
+	color: #262626;
+	text-transform: uppercase;
+	background: linear-gradient(90deg, #f9572a, #ffc905);
+	-webkit-background-clip: text;
+	-webkit-text-fill-color: transparent;
+	background-size: cover;
+	background-position: center;
+	font-family: 'Montserrat', sans-serif;
 }
 
 .not-found p,
 .error p {
-  font-size: 1.5rem;
-  font-weight: 500;
-  color: #ddd8d8;
+	font-size: 1.5rem;
+	font-weight: 500;
+	color: #ddd8d8;
 }
