+++ 25 Foodies Starting Project/components/main-header/main-header-background.module.css
@@ -1,0 +1,14 @@
+.header-background {
+	position: absolute;
+	width: 100%;
+	height: 320px;
+	top: 0;
+	left: 0;
+	z-index: -1;
+}
+
+.header-background svg {
+	display: block;
+	width: 100%;
+	height: auto;
+}
